+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is the log of notable changes to the Expo client that are developer-facing.
 - `react-native-reanimated` updated from `1.0.1` to `1.1.0`
 - `react-native-svg` updated from `9.4.0` to `9.5.1`
 - `react-native-webview` updated from `5.8.0` to `5.12.0`
+- `@react-native-community/netinfo` updated from `2.0.10` to `3.2.1`
 
 ### 🛠 Breaking changes